working on pip acquisition

+++ scripts/getpip.sh
@@ -1,6 +1,10 @@
-PYTHONVER=$(python -c "import sys; print(str(sys.version_info.major)+'.'+str(sys.version_info.minor))")
-rm -rf get-pip.py
-wget -nv https://bootstrap.pypa.io/pip/$PYTHONVER/get-pip.py
-python get-pip.py --user
+if ! command -v pip &> /dev/null; then
+    PYTHONVER=$(python -c "import sys; print(str(sys.version_info.major)+'.'+str(sys.version_info.minor))")
+    rm -rf get-pip.py
+    wget -nv https://bootstrap.pypa.io/pip/$PYTHONVER/get-pip.py
+    python get-pip.py --user   
+fi
+
 python -m pip install --user scipy
 python -m pip install --user scikit-image
+
